Removed the "Processed files" line from expected output.
I've decided to remove the "Processed all the files from <...>."
line from the expected output in the tests. It makes it hard to
get the expected output, especially if all we're doing is using
diff. If we switched from these simple shell tests to something
more powerful like Aruba then we could more easily including that
line, but I'm dropping it for now.