Remove symbolic list from `summarize_tree` test data
There are several subtle problems that can occur because fewer_files is a symbolic link into the larger tree structure of loads_o_files. We're really not saving very much by having fewer_files be a link instead of a straight up copy, so I think we should just drop the use of the link.

closes #9